+++ packages/ecosystem/material-icons/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dlightjs/material-icons
 
+## 0.10.1
+
+### Patch Changes
+
+- feat: Add @Watch and fix dep loop
+- Updated dependencies
+  - @dlightjs/decorators@0.10.1
+  - @dlightjs/dlight@0.10.1
+
 ## 0.10.0
 
 ### Minor Changes